+++ framework/innerkitsimpl/data/system_defined_form.cpp
@@ -42,8 +42,9 @@ SystemDefinedForm::SystemDefinedForm(UDType type, ValueType value) : SystemDefin
 
 int64_t SystemDefinedForm::GetSize()
 {
-    return UnifiedDataUtils::GetDetailsSize(this->details_) + sizeof(formId_) + this->formName_.size()
-           + this->bundleName_.size() + this->abilityName_.size() + this->module_.size() + GetInnerEntriesSize();
+    return static_cast<int64_t>(UnifiedDataUtils::GetDetailsSize(this->details_) + sizeof(formId_) +
+        this->formName_.size() + this->bundleName_.size() + this->abilityName_.size() + this->module_.size()) +
+        GetInnerEntriesSize();
 }
 
 int32_t SystemDefinedForm::GetFormId() const
