+++ framework/innerkitsimpl/data/system_defined_record.cpp
@@ -23,7 +23,7 @@ SystemDefinedRecord::SystemDefinedRecord() : UnifiedRecord(SYSTEM_DEFINED_RECORD
 
 int64_t SystemDefinedRecord::GetSize()
 {
-    return UnifiedDataUtils::GetDetailsSize(this->details_) + GetInnerEntriesSize();
+    return static_cast<int64_t>(UnifiedDataUtils::GetDetailsSize(this->details_)) + GetInnerEntriesSize();
 }
 
 SystemDefinedRecord::SystemDefinedRecord(UDType type, ValueType value) : UnifiedRecord(type, value)
